HA protected ledger initialization for the appendOnly [DPP-551]
* Preparation - Refactoring ParallelIndexerFactory: pulling initialization, resource factory and subscription apart

CHANGELOG_BEGIN
CHANGELOG_END
